+++ ricaun.Nuke.RevitAddin.Example/Proprieties/AssemblyInfo.cs
@@ -1,5 +1,5 @@
 ﻿using System;
 using System.Windows.Markup;
 
-//[assembly: XmlnsDefinition("http://ricaun.com/example", "ricaun.Example")]
-//[assembly: XmlnsPrefix("http://ricaun.com/example", "r")]
+[assembly: XmlnsDefinition("http://ricaun.com/example", "ricaun.Example")]
+[assembly: XmlnsPrefix("http://ricaun.com/example", "r")]