+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] / 2022-02-23
+### Changed
+- Add GetComments & GetFileDescription
+- Add ShowInformation On MainProject
+
 ## [1.1.1] / 2022-02-23
 ### New Features
 - Compile Multiple Examples with EndWith `Name`